+++ src/App.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 import { Home } from './pages/Home';
 import { Publications } from './archived/Publications';
-import { Projects } from './pages/Projects';
+import { Projects } from './pages/Projects/Projects';
 import { Research } from './archived/Research';
 import { People } from './pages/People/People';
 import { Presentations } from './archived/Presentations';
